requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
exceptiongroup==1.0.4
idna==3.4
importlib-metadata==5.1.0
iniconfig==1.1.1
Markdown==3.4.1
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.9
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
urllib3==1.26.13
zipp==3.11.0